+++ d2/src/main/java/com/linkedin/d2/balancer/properties/CanaryDistributionStrategy.java
@@ -1,5 +1,22 @@
+/*
+   Copyright (c) 2021 LinkedIn Corp.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package com.linkedin.d2.balancer.properties;
 
+import com.linkedin.util.ArgumentUtil;
 import java.util.Map;
 
 
@@ -23,6 +40,10 @@ public class CanaryDistributionStrategy
                                     Map<String, Object> targetHostsStrategyProperties,
                                     Map<String, Object> targetApplicationsStrategyProperties)
   {
+    ArgumentUtil.notNull(strategy, "strategy");
+    ArgumentUtil.notNull(percentageStrategyProperties, "percentageStrategyProperties");
+    ArgumentUtil.notNull(targetHostsStrategyProperties, "targetHostsStrategyProperties");
+    ArgumentUtil.notNull(targetApplicationsStrategyProperties, "targetApplicationsStrategyProperties");
     _strategy = strategy;
     _percentageStrategyProperties = percentageStrategyProperties;
     _targetHostsStrategyProperties = targetHostsStrategyProperties;